split averaging with export script
